# Update BLS signature ordering to avoid public key compression