--- a/vms/platformvm/warp/signature_test.go
+++ b/vms/platformvm/warp/signature_test.go
@@ -55,7 +55,7 @@ func newTestValidator() *testValidator {
 		sk:     sk,
 		vdr: &Validator{
 			PublicKey:      pk,
-			PublicKeyBytes: bls.PublicKeyToBytes(pk),
+			PublicKeyBytes: pk.Serialize(),
 			Weight:         3,
 			NodeIDs:        []ids.NodeID{nodeID},
 		},