--- a/vms/platformvm/warp/validator_test.go
+++ b/vms/platformvm/warp/validator_test.go
@@ -166,7 +166,7 @@ func TestFilterValidators(t *testing.T) {
 	pk0 := bls.PublicFromSecretKey(sk0)
 	vdr0 := &Validator{
 		PublicKey:      pk0,
-		PublicKeyBytes: bls.PublicKeyToBytes(pk0),
+		PublicKeyBytes: pk0.Serialize(),
 		Weight:         1,
 	}
 
@@ -175,7 +175,7 @@ func TestFilterValidators(t *testing.T) {
 	pk1 := bls.PublicFromSecretKey(sk1)
 	vdr1 := &Validator{
 		PublicKey:      pk1,
-		PublicKeyBytes: bls.PublicKeyToBytes(pk1),
+		PublicKeyBytes: pk1.Serialize(),
 		Weight:         2,
 	}