## Why this should be merged

Public key compression should be avoided in any code hot-paths.

## How this works

Uses the uncompressed format rather than the compressed format.

## How this was tested

CI