@aaronbuchwald 2023-04-25
Results for `BenchmarkGetCanonicalValidatorSet`

Before:

```
# github.com/ava-labs/avalanchego/vms/platformvm/warp.test
ld: warning: could not create compact unwind for _blst_sha256_block_data_order: does not use RBP or RSP based frame
goos: darwin
goarch: amd64
pkg: github.com/ava-labs/avalanchego/vms/platformvm/warp
cpu: Intel(R) Core(TM) i7-9750H CPU @ 2.60GHz
BenchmarkGetCanonicalValidatorSet/0-12  	 2944636	       402.2 ns/op	       0 B/op	       0 allocs/op
BenchmarkGetCanonicalValidatorSet/1-12  	 1000000	      1061 ns/op	     192 B/op	       5 allocs/op
BenchmarkGetCanonicalValidatorSet/10-12 	  182083	      6638 ns/op	    2340 B/op	      42 allocs/op
BenchmarkGetCanonicalValidatorSet/100-12         	   18933	     69755 ns/op	   23476 B/op	     403 allocs/op
BenchmarkGetCanonicalValidatorSet/1000-12        	    1622	    811667 ns/op	  249560 B/op	    4003 allocs/op
BenchmarkGetCanonicalValidatorSet/10000-12       	     100	  11025589 ns/op	 2380696 B/op	   40003 allocs/op
PASS
ok  	github.com/ava-labs/avalanchego/vms/platformvm/warp	11.748s
```

After:

```
ld: warning: could not create compact unwind for _blst_sha256_block_data_order: does not use RBP or RSP based frame
goos: darwin
goarch: amd64
pkg: github.com/ava-labs/avalanchego/vms/platformvm/warp
cpu: Intel(R) Core(TM) i7-9750H CPU @ 2.60GHz
BenchmarkGetCanonicalValidatorSet/0-12  	 3929092	       298.2 ns/op	       0 B/op	       0 allocs/op
BenchmarkGetCanonicalValidatorSet/1-12  	 1351539	      1023 ns/op	     288 B/op	       5 allocs/op
BenchmarkGetCanonicalValidatorSet/10-12 	  219115	      5952 ns/op	    3300 B/op	      42 allocs/op
BenchmarkGetCanonicalValidatorSet/100-12         	   21322	     60256 ns/op	   33076 B/op	     403 allocs/op
BenchmarkGetCanonicalValidatorSet/1000-12        	    1632	    839877 ns/op	  345561 B/op	    4003 allocs/op
BenchmarkGetCanonicalValidatorSet/10000-12       	     100	  10612301 ns/op	 3340696 B/op	   40003 allocs/op
PASS
ok  	github.com/ava-labs/avalanchego/vms/platformvm/warp	11.949s
```